EZP-24664: adapt Legacy engine
