+++ eZ/Publish/Core/settings/search_engines/legacy.yml
@@ -70,13 +70,8 @@ services:
         class: %ezpublish.spi.search.legacy.content_handler.class%
         arguments:
             - @ezpublish.search.legacy.gateway.content
-            - @ezpublish.persistence.legacy.content.mapper
-        lazy: true
-
-    ezpublish.spi.search.legacy.location_handler:
-        class: %ezpublish.spi.search.legacy.location_handler.class%
-        arguments:
             - @ezpublish.search.legacy.gateway.location
+            - @ezpublish.persistence.legacy.content.mapper
             - @ezpublish.persistence.legacy.location.mapper
         lazy: true
 
@@ -84,7 +79,6 @@ services:
         class: %ezpublish.spi.search.legacy.class%
         arguments:
             - @ezpublish.spi.search.legacy.content_handler
-            - @ezpublish.spi.search.legacy.location_handler
         tags:
             - {name: ezpublish.searchEngine, alias: legacy}
         lazy: true
